+++ lib/src/data/ble/facade/ble_transport_facade_impl.dart
@@ -2,6 +2,7 @@ import 'dart:async';
 
 import '../../../domain/internal/transport_message.dart';
 import '../../../domain/transport/models/transport_role.dart';
+import '../../../domain/transport/models/transport_session_disconnect_event.dart';
 import '../../../domain/transport/models/transport_session_state.dart';
 import '../../../domain/transport/transport_facade.dart';
 import '../../../domain/transport/transport_session.dart';
@@ -29,6 +30,15 @@ final class BleTransportFacadeImpl implements TransportFacade {
     _proxyServerConnectionStateStreamSubscription = _serverSession.connectionStateStream.listen(
       _proxyConnectionStateStreamController.add,
     );
+
+    _proxyDisconnectEventStreamController =
+        StreamController<TransportSessionDisconnectEvent>.broadcast();
+    _proxyClientDisconnectEventStreamSubscription = _clientSession.disconnectEventStream.listen(
+      _proxyDisconnectEventStreamController.add,
+    );
+    _proxyServerDisconnectEventStreamSubscription = _serverSession.disconnectEventStream.listen(
+      _proxyDisconnectEventStreamController.add,
+    );
   }
 
   final TransportSessionClient _transportSessionClient;
@@ -49,12 +59,23 @@ final class BleTransportFacadeImpl implements TransportFacade {
   late final StreamSubscription<TransportSessionState>
   _proxyServerConnectionStateStreamSubscription;
 
+  late final StreamController<TransportSessionDisconnectEvent>
+  _proxyDisconnectEventStreamController;
+  late final StreamSubscription<TransportSessionDisconnectEvent>
+  _proxyClientDisconnectEventStreamSubscription;
+  late final StreamSubscription<TransportSessionDisconnectEvent>
+  _proxyServerDisconnectEventStreamSubscription;
+
   @override
   Stream<TransportMessage> get messagesStream => _proxyMessagesStreamController.stream;
 
   @override
   Stream<TransportSessionState> get connectionStateStream =>
       _proxyConnectionStateStreamController.stream;
+
+  @override
+  Stream<TransportSessionDisconnectEvent> get disconnectEventStream =>
+      _proxyDisconnectEventStreamController.stream;
 
   @override
   Future<void> sendMessage(TransportMessage message) => transportSession.sendMessage(message);
@@ -96,6 +117,9 @@ final class BleTransportFacadeImpl implements TransportFacade {
     await _proxyClientConnectionStateStreamSubscription.cancel();
     await _proxyServerConnectionStateStreamSubscription.cancel();
     await _proxyConnectionStateStreamController.close();
+    await _proxyClientDisconnectEventStreamSubscription.cancel();
+    await _proxyServerDisconnectEventStreamSubscription.cancel();
+    await _proxyDisconnectEventStreamController.close();
     await _clientSession.dispose();
     await _serverSession.dispose();
   }